more fixes, and changed packet and error code numbering

+++ Autoloads/Globals.cs
@@ -2,5 +2,5 @@ namespace FourInARowBattle;
 
 public static class Globals
 {
-    public const int NAME_LENGTH_LIMIT = 32;
+    public const byte NAME_LENGTH_LIMIT = 32;
 }
